requirements.txt
# Requirements for Image Processing Project - Python 3.12 Compatible
# Tối ưu cho Windows với Python 3.12

# Core dependencies
numpy>=1.26.0
opencv-python>=4.8.0
pillow>=10.0.0

# Deep Learning (TensorFlow 2.16+ hỗ trợ Python 3.12)
tensorflow>=2.16.0

# Web Interface
streamlit>=1.28.0

# Utilities
matplotlib>=3.8.0
scikit-learn>=1.3.0

# Build tools
setuptools>=65.0.0
wheel>=0.38.0
